fix(cli): parse --threads as u16 so large core counts are reachable
The value parser was u8 while the range upper bound was the machine's
core count: on hardware with more than 255 cores the flag could never
use them all.

+++ src/main.rs
@@ -845,7 +845,7 @@ fn main() -> std::process::ExitCode {
 
     match matches.subcommand() {
         Some((subcommand, matches)) => {
-            let threads = matches.get_one::<u8>("threads").copied().unwrap_or(1) as usize;
+            let threads = matches.get_one::<u16>("threads").copied().unwrap_or(1) as usize;
 
             // Hidden diagnostic: print the runtime-derived limits and exit
             // before touching any files. Used to understand why an image was
